Added board cad files
Added KiCad files for the PCB. Custom symbols used should be included. At this stage, the schematic is finished except for the OLED modules, and the layout is approximately 25% done. Routing is not started yet. We are still shooting for a double-layer board of less than 100x150mm, which looks obtainable; but we may have to revise upward depending on the complexity of the routing.